+++ src/Ana/Ana/Ana.interface.ts
@@ -11,16 +11,25 @@
 //
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 export interface iAnaConfiguration {
-  // includeComponentLibrary?: boolean
-  // includeAtoms?: boolean
-  // includeMolecules?: boolean
-  // includeOrganisms?: boolean
-  // includePages?: boolean
-  // includeSVGLibrary?: boolean
+  /**
+   * 
+   */
   extensions?: { [key:string]: any }
 
-  // Component/Organisms/App
-  // appContainerId?: string = 'app' // index.html -> <body><div id="app"></div></body>
+  /**
+   * 
+   */
+  svgElements?: string[]
+
+  /**
+   * 
+   */
+  emptyElements?: string[]
+
+  /**
+   * index.html -> <body><div id="app"></div></body>
+   */
+  appContainerId?: string
 }
 
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
@@ -28,6 +37,7 @@ export interface AnaConfiguration {
   extensions: { [key:string]: any }
   svgElements: string[]
   emptyElements: string[]
+  appContainerId: string
 }
 
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
@@ -37,4 +47,5 @@ export const dAnaConfiguration: AnaConfiguration = {
   svgElements: ['circle','ellipse','line','polygon','polyline','rect','defs','g','marker','mask','svg','switch','symbol','desc','linearGradient','radialGradient','stop','image','path','text','use'],
   // prettier-ignore
   emptyElements: ['area','base','br','col','embed','hr','img','input','link','meta','param','source','track','wbr'],
+  appContainerId: 'app'
 }